feat: implement computed articles method and update published count handling

+++ app/Livewire/ArticleList.php
@@ -3,6 +3,7 @@
 namespace App\Livewire;
 
 use App\Models\Article;
+use Livewire\Attributes\Computed;
 use Livewire\Attributes\Title;
 use Livewire\WithPagination;
 
@@ -13,9 +14,28 @@ class ArticleList extends AdminComponent
 
     public $showPublishedOnly = false;
 
+    #[Computed]
+    public function articles()
+    {
+        $query = Article::query();
+
+        if ($this->showPublishedOnly) {
+            $query->where('published', '=', 1);
+        }
+
+        return $query->paginate(10, pageName: 'article-page');
+    }
+
     public function delete(Article $article)
     {
+        if ($this->articles->count() < 10) {
+            throw new \Exception('Nope');
+        }
+
         $article->delete();
+        unset($this->articles);
+        cache()->forget(key: 'published-count');
+        $this->dispatch('publishedCountUpdated');
     }
 
     public function showAll()
@@ -30,16 +50,8 @@ class ArticleList extends AdminComponent
         $this->resetPage(pageName: 'article-page');
     }
 
-    public function render()
-    {
-        $query = Article::query();
-
-        if ($this->showPublishedOnly) {
-            $query->where('published', '=', 1);
-        }
-
-        return view('livewire.article-list', [
-            'articles' => $query->paginate(10, pageName: 'article-page'),
-        ]);
-    }
+    // public function render()
+    // {
+    //     return view('livewire.article-list');
+    // }
 }
